requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
alembic
bcrypt==4.0.1
passlib
python-dotenv